Use preparedStatement, and close Statement and ResultSet
